use last i/o first image as thumnail
This patch changes the thumbnail image from the first to the last of the
provided series, and updates two examples accordingly. The rationale behind
this change is that scripts often build up in complexity leading the last image
to be the most impressive. While this last image could simply be listed first
to make it the thumbnail, the side effect of this is that the detailed view
shows images in counter-logical order. In the finite strain exampe, for
instance, the result of the nonlinear problem is awkwardly followed by the
intermediate result of the linear problem. This patch restores natural
left-to-right progression for this type of situation.